member 수정 및 participation 완료
project 완료

+++ models/team.js
@@ -1,12 +1,12 @@
 
 const mysql = require('mysql2/promise')
-const config = require('../configs/configs.json')
+const settings = require('../config/appsettings.local.json')
 // TODO: 팀이름 공백, 영문, 한글, 숫자, _ , -만 가능하게 검사하기
 // SQL injection!
 
 class Team {
     constructor() {
-        this._pool = mysql.createPool(config.database)
+        this._pool = mysql.createPool(settings.database)
     }
     async find(tid) {
         let connection = null
@@ -16,9 +16,9 @@ class Team {
             const [rows] = await connection.query(`SELECT * FROM team WHERE id=${tid}`)
 
             if(rows.length == 0)
-                throw "팀이 없습니다."
+                return null
 
-            return rows
+            return rows[0]
         } catch (err) {
             throw err
         } finally {
@@ -32,9 +32,6 @@ class Team {
             connection = await this._pool.getConnection()
 
             const [rows] = await connection.query(`SELECT * FROM team`)
-
-            if(rows.length == 0)
-                throw "팀이 없습니다."
 
             return rows
         } catch (err) {
@@ -52,7 +49,7 @@ class Team {
             const [rows] = await connection.query(`SELECT * FROM team WHERE owner=${owner}`)
             
             if(rows.length == 0)
-                throw "해당하는 팀이 없습니다."
+                return null
 
             return rows
         } catch (err) {
@@ -62,19 +59,61 @@ class Team {
         }
     }
 
-    async add(team) {
+    async find_by_member(member_id) {
+        
+        let connection = null
+
+        try {
+            connection = await this._pool.getConnection()
+
+            const [rows] = await connection.query(
+                `SELECT * FROM team
+                JOIN participation ON team.id=participation.team_id
+                WHERE member_id=${member_id}`
+            )
+
+            return rows
+        } catch (err) {
+            throw err
+        } finally {
+            connection?.release()
+        }
+    }
+
+    async find_by_team(team_id) {
+        
+        let connection = null
+
+        try {
+            connection = await this._pool.getConnection()
+
+            const [rows] = await connection.query(
+                `SELECT * FROM member
+                JOIN participation ON member.id=participation.member_id
+                WHERE team_id=${team_id}`
+            )
+
+            return rows
+        } catch (err) {
+            throw err
+        } finally {
+            connection?.release()
+        }
+    }
+
+    async add(team, member_id) {
         let connection = null
         try {
             connection = await this._pool.getConnection()
             // TODO: 팀 이름 정규식
 
             // 자신이 만든 팀 이름 중복여부 검사
-            const [rows] = await connection.query(`SELECT * FROM team WHERE name='${team.name}' and owner=${team.owner}`)
+            const [rows] = await connection.query(`SELECT * FROM team WHERE name='${team.name}' and owner=${member_id}`)
             
             if(rows.length > 0)
-                throw "팀이름이 중복됩니다."
+                return false
 
-            const [result] = await connection.query(`INSERT INTO team(name,owner) values('${team.name}',${team.owner})`)
+            const [result] = await connection.query(`INSERT INTO team(name,owner) values('${team.name}',${member_id})`)
 
             return result.insertId
         } catch (err) {
@@ -84,29 +123,23 @@ class Team {
         }
     }
 
-    async modify(mid, team) {
+    async update(mid, team) {
         let connection = null
         try {
             connection = await this._pool.getConnection()
 
-            // 권한 검사
-            const [rows1] = await connection.query(`SELECT * FROM team WHERE id=${team.id}`)
-            
-            if(rows1.length == 0)
-                throw "팀이 존재하지 않습니다."
-            if(rows1[0].owner != mid)
-                throw "권한이 없습니다."
-
             // 팀 이름 중복 검사
-            const [rows2] = await connection.query(
+            const [exists] = await connection.query(
                 `SELECT * FROM team
                 WHERE NOT id=${team.id} AND name='${team.name}' AND owner=${mid}`
             )
 
-            if(rows2.length > 0)
-                throw "팀 이름이 중복됩니다."
+            if(exists.length > 0)
+                return null
 
-            await this._pool.query(`UPDATE team SET name='${team.name}',owner='${team.owner}' WHERE id=${team.id}`)
+            const [result] = await this._pool.query(`UPDATE team SET name='${team.name}',owner='${team.owner}' WHERE id=${team.id}`)
+
+            return result.affectedRows > 0
 
         } catch (err) {
             throw err
@@ -116,29 +149,21 @@ class Team {
     }
 
     async remove(id, mid) {
-        let connection = await this._pool.getConnection()
+        let connection = null
         
         try {
             connection = await this._pool.getConnection()
 
-            const [rows] = await connection.query(
-                `SELECT * FROM team WHERE id=${id} AND owner=${mid}`
-            )
-
-            if(rows.length == 0)
-                throw "권한이 없습니다."
-            
             const [result] = await connection.query(
-                `DELETE FROM team WHERE id=${id}`
+                `DELETE FROM team WHERE id=${id} AND owner=${mid}`
             )
+            return result.affectedRows > 0
         } catch (err) {
             throw err
         } finally {
             connection?.release()
         }
     }
-
-    
 }
 
 module.exports = Team
